feat: Reintroduce observability for bullmq
