+++ control-plane/src/modules/queues/external-tool-call.ts
@@ -1,7 +1,7 @@
-import { BaseMessage } from "../sqs";
 import { handleExternalCall } from "../jobs/external";
 import { createQueue } from "./core";
 import { QueueNames } from "./core";
+import { BaseMessage } from "./observability";
 
 interface ExternalToolCallMessage extends BaseMessage {
   jobId: string;
